+++ tests/factories/populate_database.py
@@ -10,20 +10,20 @@ from tests.factories.marketplace import ServiceFactory, UserFactory
 
 
 def populate_users_and_services(
-    common_services_number,
-    no_one_services_number,
-    users_number,
+    common_services_no,
+    unordered_services_no,
+    total_users,
     k_common_services_min,
     k_common_services_max,
     verbose=False,
 ):
-    """Populate database with users and their accessed (or not) services"""
+    """Populate database with users and their ordered (or not) services"""
 
-    _no_one_services = [
+    _unordered_services = [
         ServiceFactory()
         for _ in trange(
-            no_one_services_number,
-            desc="No one services creating...",
+            unordered_services_no,
+            desc="Creating unordered services...",
             disable=(not verbose),
         )
     ]
@@ -31,14 +31,14 @@ def populate_users_and_services(
     common_services = [
         ServiceFactory()
         for _ in trange(
-            common_services_number,
-            desc="Common services creating...",
+            common_services_no,
+            desc="Creating common services ...",
             disable=(not verbose),
         )
     ]
 
-    for _ in trange(users_number, desc="Users creating...", disable=(not verbose)):
+    for _ in trange(total_users, desc="Creating users...", disable=(not verbose)):
         k = random.randint(k_common_services_min, k_common_services_max)
         accessed_services = random.sample(common_services, k=k)
-        user = UserFactory(accessed_services=accessed_services)
+        user = UserFactory(accessed_services=accessed_services, synthetic=True)
         user.save()